// data/github/current/ui/packages/react-sandbox/routes/__generated__/IssueSubIssues_InnerFragment.graphql.ts
export type IssueState = "CLOSED" | "OPEN" | "%future added value";
import type { FragmentRefs } from "relay-runtime";
export type IssueSubIssues_InnerFragment$data = {
  readonly assignees: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly avatarUrl: string;
        readonly id: string;
        readonly login: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly closedByPullRequestsReferences: {
    readonly totalCount: number;
  } | null | undefined;
  readonly id: string;
  readonly labels: {
    readonly nodes: ReadonlyArray<{
      readonly color: string;
      readonly id: string;
      readonly name: string;
    } | null | undefined> | null | undefined;
  } | null | undefined;
  readonly state: IssueState;
  readonly subIssuesSummary: {
    readonly completed: number;
    readonly percentCompleted: number;
    readonly total: number;
  };
  readonly title: string;
  readonly titleHTML: string;
  readonly url: string;
  readonly " $fragmentType": "IssueSubIssues_InnerFragment";
};
export type IssueSubIssues_InnerFragment$key = {
  readonly " $data"?: IssueSubIssues_InnerFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"IssueSubIssues_InnerFragment">;
};